Renamed platform chromium to chrome
