+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "focused-youtube",
-  "version": "3.0.0",
+  "version": "3.0",
   "scripts": {
-    "build": "mode=${npm_config_mode:-chromium} && vite build --config vite.popup.config.js --mode $mode && vite build --config vite.background.config.js --mode $mode && vite build --config vite.welcome.config.js --mode $mode && vite build --config vite.content_script.config.js --mode $mode && mkdir -p dist_$mode && rsync -av output_background_script/* dist_$mode && rsync -av output_popup/* dist_$mode && rsync -av output_welcome/* dist_$mode && rsync -av output_content_script/* dist_$mode && rm dist_$mode/index.html && node scripts/build.js --mode $mode",
-    "build:all": "npm run build:chromium && npm run build:firefox && npm run zip:all",
-    "build:chromium": "npm run build --mode=chromium",
+    "build": "mode=${npm_config_mode:-chrome} && vite build --config vite.popup.config.js --mode $mode && vite build --config vite.background.config.js --mode $mode && vite build --config vite.welcome.config.js --mode $mode && vite build --config vite.content_script.config.js --mode $mode && mkdir -p dist_$mode && rsync -av output_background_script/* dist_$mode && rsync -av output_popup/* dist_$mode && rsync -av output_welcome/* dist_$mode && rsync -av output_content_script/* dist_$mode && rm dist_$mode/index.html && node scripts/build.js --mode $mode",
+    "build:all": "npm run build:chrome && npm run build:firefox && npm run zip:all",
+    "build:chrome": "npm run build --mode=chrome",
     "build:firefox": "npm run build --mode=firefox",
     "build:edge": "npm run build --mode=edge",
-    "zip:all": "zip -r \"focused-youtube-chromium.zip\" dist_chromium && cd dist_firefox && zip -r \"../focused-youtube-firefox.zip\" .",
+    "zip:all": "zip -r \"focused-youtube-chrome.zip\" dist_chrome && cd dist_firefox && zip -r \"../focused-youtube-firefox.zip\" .",
     "dev": "nodemon --ext js,vue,css,json,svg,png,html,ttf --watch 'src/**' --watch 'public/**' --watch 'platforms/**' --exec 'npm run build && npm run zip:all'",
     "lint": "web-ext lint -s dist_firefox && eslint --ext .js,.vue --ignore-pattern .gitignore --fix src",
     "format": "prettier .  --write"
