+++ platforms/chrome/manifest.json
