--- src/helpers/browser-api.js
@@ -1,5 +1,0 @@
-/* global browser */
-
-const browserAPI = typeof browser !== "undefined" ? browser : chrome
-
-export default browserAPI
