+++ src/helpers/browser.js
@@ -1,0 +1,5 @@
+/* global browser, __BROWSER_PLATFORM__ */
+
+export const browserAPI = typeof browser !== "undefined" ? browser : chrome
+
+export const browserPlatform = __BROWSER_PLATFORM__
