+++ vite.background.config.js
@@ -6,7 +6,7 @@ import { fileURLToPath } from "url"
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
-export default defineConfig({
+export default defineConfig(({ mode }) => ({
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
@@ -18,6 +18,9 @@ export default defineConfig({
     vue(),
     eslint()
   ],
+  define: {
+    "__BROWSER_PLATFORM__": JSON.stringify(mode)
+  },
   build: {
     rollupOptions: {
       input: "src/background.js",
@@ -28,4 +31,4 @@ export default defineConfig({
       }
     }
   }
-})
+}))
